Fix XRefs
This PR fixes all the cross-reference, with the exception
of a few annotated XRefs which are either optional (not
in the original but useful), to external documents, or
inside EBNF rules and need to be discussed first.

All cross-references have been thoroughly checked against
the scanned images of the original document, e.g. checking
the mentioned page number in the original xref.

Due to some sections sharing a same title, custom section IDs
were added to enable XRefs targeting, adopting an IDs naming
convention which is fully documented in the README file, which
also provides lookup table for all the custom IDs added.

All section titles have been normalized to title casing
(Chicago Manual of Style), including the uppercase titles
except for titles which are CLI options, which were kept
upper case (as in original) to make it clear they refer
to options.